requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
configparser==4.0.2
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
Markdown==3.1.1
mysqlclient==1.4.4
psycopg2==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
requests==2.22.0
requests-file==1.4.3
rsa==4.0
six==1.12.0
sqlparse==0.3.0
tldextract==2.2.2
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
